+++ src/index.ts
@@ -1,4 +1,2 @@
 export { default as schema } from "./schema";
-export { default as types } from "./types";
-// export * from "./binding";
-// export * from "./nexus";
+export { default as client } from "./client";
