+++ src/schema.ts
@@ -1,6 +1,13 @@
 import * as path from "path";
+import * as glob from "glob";
 import { makeSchema } from "@nexus/schema";
-import types from "./types";
+
+let types = {};
+const files = glob.sync(path.resolve(__dirname, "./graphql/**/*.ts"));
+files.forEach((file) => {
+  const module = require(path.resolve(file));
+  types = { ...types, ...module };
+});
 
 const schema = makeSchema({
   types,
