+++ src/server.ts
@@ -1,17 +1,8 @@
 import { ApolloServer } from "apollo-server";
-import Stripe from "stripe";
 import schema from "./schema";
-
-const stripe = new Stripe(process.env.STRIPE_SECRET_KEY || "", {
-  apiVersion: "2020-03-02",
-});
-
-const server = new ApolloServer({
-  schema,
-  context: {
-    stripe,
-  },
-});
+import stripe from "./stripe";
+import client from "./client";
+const server = new ApolloServer({ schema, context: { stripe } });
 
 const port = process.env.PORT || 4000;
