+++ src/stripe.ts
@@ -1,0 +1,7 @@
+import Stripe from "stripe";
+
+const stripe = new Stripe(process.env.STRIPE_SECRET_KEY || "", {
+  apiVersion: "2020-03-02",
+});
+
+export default stripe;
